Get rid of SDT runtime dep
