+++ calico/src/main/scala/calico/html/keys/Prop.scala
@@ -143,3 +143,83 @@ trait StylePropModifiers[F[_]](using F: Async[F]):
         os.fold(e.removeAttribute("style"))(e.style = _)
         ()
       })(_.values)
+
+final class ClassProp[F[_]] private[calico]
+    extends HtmlProp[F, List[String], String](
+      "className",
+      Codec.whitespaceSeparatedStringsCodec
+    ):
+  import ClassProp.*
+
+  inline def :=(cls: String): SingleConstantModifier =
+    SingleConstantModifier(cls)
+
+object ClassProp:
+  final class SingleConstantModifier(val cls: String)
+
+trait ClassPropModifiers[F[_]](using F: Async[F]):
+  import ClassProp.*
+  inline given forConstantClassProp[N]: Modifier[F, N, SingleConstantModifier] =
+    _forConstantClassProp.asInstanceOf[Modifier[F, N, SingleConstantModifier]]
+  private val _forConstantClassProp: Modifier[F, Any, SingleConstantModifier] =
+    (m, n) => Resource.eval(F.delay(n.asInstanceOf[js.Dictionary[String]]("className") = m.cls))
+
+final class DataProp[F[_]] private[calico] (name: String):
+  import DataProp.*
+
+  inline def :=(v: String): ConstantModifier =
+    ConstantModifier(name, v)
+
+  inline def <--(vs: Signal[F, String]): SignalModifier[F] =
+    SignalModifier(name, vs)
+
+  inline def <--(vs: Signal[F, Option[String]]): OptionSignalModifier[F] =
+    OptionSignalModifier(name, vs)
+
+object DataProp:
+  final class ConstantModifier(
+      val name: String,
+      val value: String
+  )
+
+  final class SignalModifier[F[_]](
+      val name: String,
+      val values: Signal[F, String]
+  )
+
+  final class OptionSignalModifier[F[_]](
+      val name: String,
+      val values: Signal[F, Option[String]]
+  )
+
+trait DataPropModifiers[F[_]](using F: Async[F]):
+  import DataProp.*
+
+  private inline def setDataProp[N](node: N, value: String, name: String) =
+    F.delay(node.asInstanceOf[dom.HTMLElement].dataset(name) = value)
+
+  inline given forConstantDataProp[N <: fs2.dom.HtmlElement[F]]
+      : Modifier[F, N, ConstantModifier] =
+    _forConstantDataProp.asInstanceOf[Modifier[F, N, ConstantModifier]]
+
+  private val _forConstantDataProp: Modifier[F, fs2.dom.HtmlElement[F], ConstantModifier] =
+    (m, n) => Resource.eval(setDataProp(n, m.value, m.name))
+
+  inline given forSignalDataProp[N]: Modifier[F, N, SignalModifier[F]] =
+    _forSignalDataProp.asInstanceOf[Modifier[F, N, SignalModifier[F]]]
+
+  private val _forSignalDataProp: Modifier[F, Any, SignalModifier[F]] =
+    Modifier.forSignal[F, Any, SignalModifier[F], String]((any, sm, s) =>
+      setDataProp(any, s, sm.name))(_.values)
+
+  inline given forOptionSignalDataProp[N]: Modifier[F, N, OptionSignalModifier[F]] =
+    _forOptionSignalDataProp.asInstanceOf[Modifier[F, N, OptionSignalModifier[F]]]
+
+  private val _forOptionSignalDataProp: Modifier[F, Any, OptionSignalModifier[F]] =
+    Modifier.forSignal[F, Any, OptionSignalModifier[F], Option[String]]((any, osm, os) =>
+      F.delay {
+        val e = any.asInstanceOf[dom.HTMLElement]
+        os.fold(e.dataset -= osm.name)(v => e.dataset(osm.name) = v)
+        ()
+      })(_.values)
+
